requirements.txt
openai 
aiogram 
pandas 
tiktoken 
scipy 
nest_asyncio 
python-dotenv